requirements.txt
pygobject
playsound
pycoingecko
apscheduler
termcolor
daemonize
python-binance